Last update

+++ Sports-Playlist-Server/Controllers/PlaylistController.cs
@@ -3,10 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Sports_Playlist_Server.DTOs;
 using Sports_Playlist_Server.Extensions;
 using Sports_Playlist_Server.Interfaces;
+using Sports_Playlist_Server.Models;
 
 namespace Sports_Playlist_Server.Controllers
 {
@@ -16,33 +19,42 @@ namespace Sports_Playlist_Server.Controllers
     public class PlaylistController : ControllerBase
     {
         private readonly IPlaylistRepository _playlistRepository;
+        private readonly UserManager<User> _userManager;
 
-        public PlaylistController(IPlaylistRepository playlistRepository)
+        public PlaylistController(IPlaylistRepository playlistRepository, UserManager<User> userManager)
         {
             _playlistRepository = playlistRepository;
+                        _userManager = userManager;
         }
 
         [HttpPost]
         public async Task<IActionResult> AddMatchToUserPlaylist([FromBody] int matchId)
         {
             var currentLoggedInUserId = HttpContext.User.GetUserId();
+            var currentLoggedInUser = await _userManager.Users
+                    .SingleOrDefaultAsync(x => x.Id == currentLoggedInUserId.ToString());
+
+            if (currentLoggedInUser == null)
+            {
+                return NotFound(new { Error = "Can't add match to undefined user playlist." });
+            }
 
             try
             {
                 await _playlistRepository.AddPlaylistToUser(currentLoggedInUserId.ToString(), matchId);
-                return Ok(new {Message = "Match successfully added to playlist."});
+                return Ok(new { Message = "Match successfully added to playlist." });
             }
             catch (InvalidOperationException ex)
             {
-                return BadRequest(new {Error = ex.Message});
+                return BadRequest(new { Error = ex.Message });
             }
             catch (KeyNotFoundException ex)
             {
-                return NotFound(new {Error = ex.Message});
+                return NotFound(new { Error = ex.Message });
             }
             catch (Exception ex)
             {
-                return StatusCode(500, new {Error = "An unexpected error occurred: " + ex.Message});
+                return StatusCode(500, new { Error = "An unexpected error occurred: " + ex.Message });
             }
         }
 
@@ -50,23 +62,26 @@ namespace Sports_Playlist_Server.Controllers
         public async Task<IActionResult> DeleteMatchFromUserPlaylist(int matchId)
         {
             var currentLoggedInUserId = HttpContext.User.GetUserId();
+            var currentLoggedInUser = await _userManager.Users
+                    .SingleOrDefaultAsync(x => x.Id == currentLoggedInUserId.ToString());
+
+            if (currentLoggedInUser == null)
+            {
+                return NotFound(new { Error = "Can't delete match to undefined user playlist." });
+            }
 
             try
             {
                 await _playlistRepository.DeleteMatchFromPlayList(currentLoggedInUserId.ToString(), matchId);
-                return Ok(new {Message = "Match successfully removed from playlist."});
+                return Ok(new { Message = "Match successfully removed from playlist." });
             }
             catch (InvalidOperationException ex)
             {
-                return BadRequest(new {Error = ex.Message});
-            }
-            catch (KeyNotFoundException ex)
-            {
-                return NotFound(new {Error = ex.Message});
+                return BadRequest(new { Error = ex.Message });
             }
             catch (Exception ex)
             {
-                return StatusCode(500,new {Error = "An unexpected error occurred: " + ex.Message});
+                return StatusCode(500, new { Error = "An unexpected error occurred: " + ex.Message });
             }
         }
     }
